Login path compatibility with GameServer.exe. Support for official server software. - Send player ID and file ID in WELCOME_AGREE packets during login - Handle WELCOME_MSG reply of '3', which means the character login was rejected

+++ EOLib/Domain/Login/PlayerInfoRepository.cs
@@ -8,7 +8,7 @@ namespace EOLib.Domain.Login
 
         string PlayerPassword { get; set; }
 
-        int PlayerID { get; set; }
+        short PlayerID { get; set; }
 
         bool IsFirstTimePlayer { get; set; }
 
@@ -21,7 +21,7 @@ namespace EOLib.Domain.Login
 
         string PlayerPassword { get; }
 
-        int PlayerID { get; }
+        short PlayerID { get; }
 
         bool IsFirstTimePlayer { get; }
 
@@ -35,7 +35,7 @@ namespace EOLib.Domain.Login
 
         public string PlayerPassword { get; set; }
 
-        public int PlayerID { get; set; }
+        public short PlayerID { get; set; }
 
         public bool IsFirstTimePlayer { get; set; }
